SPACE JESUS: THE RETURN OF JUDGEMENT V.1.0

+++ inc/asteroid.h
@@ -5,10 +5,10 @@
 #define _ASTEROID_H_
 
 struct asteroid_t {
-    int8_t x, y, vx, vy, tempx, tempy, tempf, active, G;
+    int32_t x, y, vx, vy, tempx, tempy, tempf, active, G;
 };
 
-void init_asteroid(struct asteroid_t * as, int8_t x, int8_t y, int8_t vx, int8_t vy);
+void init_asteroid(struct asteroid_t * as, int32_t x, int32_t y, int32_t vx, int32_t vy);
 void draw_asteroid(struct asteroid_t * as);
 void delete_asteroid(struct asteroid_t * as);
 void move_asteroid(struct asteroid_t * as);
